Use composer to load code.
- all code moved from main file into separate classes
